Use client input method for entering text.
The VNC protocol specifies key events as using X11 keysyms.
You may think that this means that servers will:
  * use the server's input method (no),
  * at least use the server's layout (nope),
  * at the very least track the state of shift key (not even this).

In reality, it is more appropriate to divide the key events into
two classes.

The first class is special and modifier keys.
These are mirrored on the server directly: a press translates
to a press, and is available for applications to catch and change
their state appropriately.

The second class is alphanumeric keys and otherwise any sequence
of keys that inputs a loosely defined "text".
These are not actually handled as keys. Instead, they are treated
as a sequence of Unicode characters, and for every character
a key press and a key release event is sent, in effect using
key events as a mere transport mechanism.

The classic way to transfer Unicode over VNC was to convert it
back to X11 keysyms, which are similar but not quite like Unicode
in many ways, but today X11 keysyms actually embed the entire
Unicode space at 0x01000000. This will probably break some
ancient servers, but it worked on everything I tried.

Note that key sequences containing Ctrl are treated specially.
Not only many applications (famously Firefox for almost fifteen
years) had problems recognizing shortcuts in non-QWERTY layouts,
but also SDL does not emit TextInput events while Ctrl is pressed
anyway.